# swagger.yaml.yaml
openapi: 3.0.0
info:
  title: Combined API Specification
  version: 1.0.0
paths:
  /cluster:
    post:
      summary: Create new cluster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                isGlobal:
                  type: boolean
                isOrganizationGlobal:
                  type: boolean
                aws:
                  type: object
                  properties:
                    region:
                      type: string
                    credentialId:
                      type: string
                agent:
                  type: object
                  properties:
                    cpu:
                      type: string
                    memory:
                      type: string
                    replicas:
                      type: integer
                      format: int32
                type:
                  type: string
                metadata:
                  type: object
                description:
                  type: string
              required:
                - name
                - isGlobal
                - isOrganizationGlobal
                - type
      responses:
        '200':
          description: Cluster created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cluster:
                    $ref: '#/components/schemas/ClusterEntity'
                  applyUrl:
                    type: string
                  shouldAgentUpdate:
                    type: boolean
                  latestAgentVersion:
                    type: string
                  pluginMetadata:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClusterPluginEntity'
  /cluster/{id}:
    put:
      summary: Update a cluster
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: header
          name: x-cluster-id
          schema:
            type: string
          required: true
        - in: header
          name: x-project-id
          schema:
            type: string
          required: true
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
        - in: header
          name: traceId
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                isGlobal:
                  type: boolean
                isOrganizationGlobal:
                  type: boolean
                aws:
                  type: object
                  properties:
                    region:
                      type: string
                    credentialId:
                      type: string
                agent:
                  type: object
                  properties:
                    cpu:
                      type: string
                    memory:
                      type: string
                    replicas:
                      type: integer
                      format: int32
                type:
                  type: string
                metadata:
                  type: object
                description:
                  type: string
              required:
                - name
                - isGlobal
                - isOrganizationGlobal
                - type
      responses:
        '200':
          description: Cluster updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cluster:
                    $ref: '#/components/schemas/ClusterEntity'
                  applyUrl:
                    type: string
                  shouldAgentUpdate:
                    type: boolean
                  latestAgentVersion:
                    type: string
                  pluginMetadata:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClusterPluginEntity'
    get:
      summary: Get a cluster by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: header
          name: x-cluster-id
          schema:
            type: string
          required: true
        - in: header
          name: x-project-id
          schema:
            type: string
          required: true
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
        - in: header
          name: traceId
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Cluster retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cluster:
                    $ref: '#/components/schemas/ClusterEntity'
                  applyUrl:
                    type: string
                  shouldAgentUpdate:
                    type: boolean
                  latestAgentVersion:
                    type: string
                  pluginMetadata:
                    type: array
                    items:
                      $ref: '#/components/schemas/ClusterPluginEntity'
    delete:
      summary: Delete a cluster by ID
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: header
          name: x-cluster-id
          schema:
            type: string
          required: true
        - in: header
          name: x-project-id
          schema:
            type: string
          required: true
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Cluster deleted successfully
  /cluster/soft:
    delete:
      summary: Soft delete clusters
      parameters:
        - in: header
          name: x-cluster-id
          schema:
            type: string
        - in: header
          name: x-project-id
          schema:
            type: string
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Clusters soft deleted successfully
...
  /cluster/{id}/update-envs:
    put:
      summary: Update envs for a cluster
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: header
          name: x-cluster-id
          schema:
            type: string
          required: true
        - in: header
          name: x-project-id
          schema:
            type: string
          required: true
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                addEnvs:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      id:
                        type: string
                    required:
                      - name
                      - id
                removeEnvs:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      id:
                        type: string
                    required:
                      - name
                      - id
      responses:
        '200':
          description: Envs updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  cluster:
                    $ref: '#/components/schemas/ClusterEntity'
                  applyUrl:
                    type: string
  /cluster/{id}/plugin-metadata:
    post:
      summary: Create a plugin metadata
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: header
          name: x-cluster-id
          schema:
            type: string
          required: true
        - in: header
          name: x-project-id
          schema:
            type: string
          required: true
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterPluginEntity'
      responses:
        '200':
          description: Plugin metadata created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  plugin:
                    $ref: '#/components/schemas/ClusterPluginEntity'
...
  /cluster/{id}/plugin-metadata/{pluginId}:
    put:
      summary: Update a plugin metadata
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: path
          name: pluginId
          schema:
            type: string
          required: true
        - in: header
          name: x-cluster-id
          schema:
            type: string
          required: true
        - in: header
          name: x-project-id
          schema:
            type: string
          required: true
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClusterPluginEntity'
      responses:
        '200':
          description: Plugin metadata updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  plugin:
                    $ref: '#/components/schemas/ClusterPluginEntity'
    delete:
      summary: Delete a plugin metadata
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: path
          name: pluginId
          schema:
            type: string
          required: true
        - in: header
          name: x-cluster-id
          schema:
            type: string
          required: true
        - in: header
          name: x-project-id
          schema:
            type: string
          required: true
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Plugin metadata deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  plugin:
                    $ref: '#/components/schemas/ClusterPluginEntity'
  /cluster/:
    get:
      summary: Get all clusters
      parameters:
        - in: header
          name: x-project-id
          schema:
            type: string
          required: true
        - in: header
          name: x-organization-id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Clusters retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClusterEntity'
...
components:
  schemas:
    ClusterEntity:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        organizationId:
          type: string
        name:
          type: string
        clusterVersion:
          type: string
        agentOperatorVersion:
          type: string
        clusterSecretKey:
          type: string
        clusterJwtKeyRef:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        connected:
          type: boolean
        metadata:
          type: object
        lastSeen:
          type: string
          format: date-time
        agentConnectTime:
          type: string
          format: date-time
        isGlobal:
          type: boolean
        isOrganizationGlobal:
          type: boolean
        secretStore:
          type: string
        aws:
          type: object
          properties:
            region:
              type: string
            credentialId:
              type: string
        agent:
          type: object
          properties:
            cpu:
              type: string
            memory:
              type: string
            replicas:
              type: integer
              format: int32
        type:
          type: string
        envs:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
            required:
              - id
              - name
        description:
          type: string
        deletedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - organizationId
        - name
        - createdAt
        - updatedAt
        - connected
        - isGlobal
        - isOrganizationGlobal
        - secretStore
        - type
    ClusterPluginEntity:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        organizationId:
          type: string
        clusterId:
          type: string
        metadata:
          type: object
        pluginName:
          type: string
        pluginType:
          type: string
        url:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - organizationId
        - clusterId
        - pluginName
        - pluginType
...